Refactor : RoomDetail 실시간 타이머 코드 hook으로 분리
